fix(sim): native f32 state representation for capture-aligned channels
The weapon table's transcribed decimal literals and the projectile
life_timer init were stored as f64, so raw state channels (reload
timers, cooldowns) differed from the f32 values a native capture
records on nearly every tick - the dominant noise source in strict
trace diffs. The table floats are now quantized to f32 (exposing a
native off-by-one: regression-bullets pistol reload costs 759 xp, not
760), reload timer stores spill once like the native f32 field, and
projectile life_timer initializes to f32(0.4). The zig runtime's
bonus-drop kill burst now carries the dedicated native 0x0041FBxx rng
caller tags instead of the generic effect_spawn_burst tags, and its
projectile spawn velocity uses the native (cos, sin) components instead
of the heading-rotated direction. Python and zig verify payloads remain
byte-identical on the kill-replay differential.

+++ src/crimson/projectiles/runtime/projectile_pool.py
@@ -141,7 +141,8 @@ class ProjectilePool:
             float(f32(math.sin(float(angle_f32)) * 1.5)),
         )
         entry.type_id = type_id
-        entry.life_timer = 0.4
+        # Native stores the f32 literal 0.4.
+        entry.life_timer = float(f32(0.4))
         entry.reserved = 0.0
         entry.speed_scale = 1.0
         entry.travel_budget = float(travel_budget)
